+++ webapp/src/components/MarkerColors/index.css
@@ -1,4 +1,4 @@
-@import '../vars.css';
+@import '../../vars.css';
 
 .marker_colors__container {
     display: flex;
