+++ webapp/src/utils/markers.js
@@ -1,10 +1,5 @@
-import React from 'react';
-import { Map, TileLayer, Marker } from 'react-leaflet';
 import { Icon } from 'leaflet';
-import PropTypes from 'prop-types';
 import isSubset from 'is-subset';
-
-import './MapComponent.css';
 
 const redIcon = new Icon({
   iconUrl: 'https://cdn.rawgit.com/pointhi/leaflet-color-markers/master/img/marker-icon-2x-red.png',
@@ -51,7 +46,7 @@ const greyIcon = new Icon({
   shadowSize: [41, 41],
 });
 
-function findCorrectMarker(object, workOrders) {
+export default function findCorrectMarker(object, workOrders) {
   const obj = workOrders.filter((item) => isSubset(object, item));
   if (obj.length !== 0) {
     return greenIcon;
@@ -67,45 +62,3 @@ function findCorrectMarker(object, workOrders) {
   }
   return redIcon;
 }
-
-
-const MapComponent = ({ objects, onMarkerClick, workOrders }) => (
-  <Map className="map" center={[63.42, 10.39]} zoom={12}>
-    <TileLayer
-      attribution='&amp;copy <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-      url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-    />
-    {objects.map((object) => (
-      <Marker
-        key={object._id}
-        position={object.coordinates}
-        icon={findCorrectMarker(object, workOrders)}
-        onClick={() => {
-          onMarkerClick(object._id);
-        }}
-      />
-    ))}
-  </Map>
-);
-
-MapComponent.propTypes = {
-  objects: PropTypes.arrayOf(
-    PropTypes.shape({
-      _id: PropTypes.string.isRequired,
-      coordinates: PropTypes.array,
-      type: PropTypes.string,
-      priority: PropTypes.number,
-      __v: PropTypes.number,
-    }),
-  ),
-  onMarkerClick: PropTypes.func,
-  workOrders: PropTypes.arrayOf(PropTypes.object),
-};
-
-MapComponent.defaultProps = {
-  objects: [],
-  onMarkerClick: null,
-  workOrders: [],
-};
-
-export default MapComponent;
